+++ packages/core/src/index.ts
@@ -1,4 +1,4 @@
-export const VERSION = "0.6.9";
+export const VERSION = "0.6.10";
 
 export type {
 	CreateComponentArgs,